requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Authlib==1.2.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cmake==3.26.3
cryptography==40.0.2
dataclasses-json==0.5.7
decorator==5.1.1
dnspython==2.3.0
faiss-cpu==1.7.3
filelock==3.11.0
frozenlist==1.3.3
gptcache==0.1.11
greenlet==2.0.2
huggingface-hub==0.13.4
idna==3.4
Jinja2==3.1.2
langchain==0.0.141
lit==16.0.1
loguru==0.7.0
lxml==4.9.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==0.27.4
openapi-schema-pydantic==1.2.4
packaging==23.1
Pillow==9.5.0
pinecone-client==2.2.1
pycparser==2.21
pydantic==1.10.7
pypdf==3.8.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
six==1.16.0
SQLAlchemy==1.4.47
sympy==1.11.1
tenacity==8.2.2
tiktoken==0.3.3
tokenizers==0.13.3
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.28.1
triton==2.0.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
validators==0.20.0
weaviate-client==3.15.6
yarl==1.8.2
